+++ StepSequencerComponent.py
@@ -1,5 +1,7 @@
 import Live
-from _Framework.ControlSurfaceComponent import ControlSurfaceComponent
+
+from .LoopSelectorComponent import LoopSelectorComponent
+from .NoteSelectorComponent import NoteSelectorComponent
 from _Framework.CompoundComponent import CompoundComponent
 from _Framework.ButtonElement import ButtonElement
 from _Framework.Util import find_if
@@ -16,7 +18,6 @@ try:
     exec("from .Settings import Settings")
 except ImportError:
     exec("from .Settings import *")
-
 # quantization button colours. this must remain of length 4.
 QUANTIZATION_MAP = [1, 0.5, 0.25, 0.125]  # 1/4 1/8 1/16 1/32
 QUANTIZATION_NAMES = ["1/4", "1/8", "1/16", "1/32"]
@@ -27,663 +28,7 @@ STEPSEQ_MODE_SCALE_EDIT = 10
 
 LONG_BUTTON_PRESS = 1.0
 
-#Allows to note selection and navigation through note groups and pages
-class NoteSelectorComponent(ControlSurfaceComponent):
-
-    def __init__(self, step_sequencer, offset_buttons, control_surface):
-        self._step_sequencer = step_sequencer
-        self._control_surface = control_surface
-        ControlSurfaceComponent.__init__(self)
-        self.set_enabled(False)
-
-        self._clip = None
-        self._track = None
-        self._notes = None
-        self._playhead = None #index of the playing position
-
-        #First note for the 16 offset buttons
-        self._root_note = 36
-        #Relative offset selected button [0-15]
-        self._offset = 0
-        self._key = 0
-        self._scale = [0, 2, 4, 5, 7, 9, 11, 12]#Major
-        self._force = True
-
-        self._up_button = None
-        self._down_button = None
-
-        self._offset_buttons = offset_buttons
-        self._enable_offset_button = True # self._mode == STEPSEQ_MODE_NORMAL
-        # cache to optimize display for offset buttons minimizing MIDI traffic
-        self._cache = [-1, -1, -1, -1, 
-                        -1, -1, -1, -1, 
-                        -1, -1, -1, -1, 
-                        -1, -1, -1, -1]
-        self._was_velocity_shifted = False
-        for button in self._offset_buttons:
-            assert isinstance(button, ButtonElement)
-            button.remove_value_listener(self.note_offset_button_value)
-            button.add_value_listener(self.note_offset_button_value, identify_sender=True)
-
-    def set_clip(self, clip):
-        self._clip = clip
-
-    def set_note_cache(self, note_cache):
-        self._note_cache = note_cache
-
-    def set_playhead(self, playhead):
-        self._playhead = playhead
-        self._update_matrix()
-
-    @property
-    def _is_mute_shifted(self):
-        return self._step_sequencer._is_mute_shifted
-
-    @property
-    def _is_velocity_shifted(self):
-        return self._step_sequencer._is_velocity_shifted
-
-    @property
-    def _drum_group_device(self):
-        return self._step_sequencer._drum_group_device
-
-    @property
-    def is_drumrack(self):
-        return self._step_sequencer._scale_selector.is_drumrack and self._drum_group_device != None
-
-    @property
-    def is_chromatic(self):
-        return self._step_sequencer._scale_selector.is_chromatic
-
-    @property
-    def is_diatonic(self):
-        return self._step_sequencer._scale_selector.is_diatonic
-        
-    
-    # Updates the DOWN button light OK    
-    def _update_down_button(self):
-        if self.is_enabled():
-            if self._down_button != None:
-                if self._clip == None:
-                    self._down_button.set_on_off_values("DefaultButton.Disabled", "DefaultButton.Disabled")
-                    self._down_button.turn_off()
-                else:
-                    self._down_button.set_on_off_values("StepSequencer.Octave.On", "StepSequencer.Octave.Off")
-                    if self.should_scroll():
-                        if self.can_scroll_down():
-                            self._down_button.turn_on()
-                        else:
-                            self._down_button.turn_off()
-                    else:
-                        if self.can_page_down():
-                            self._down_button.turn_on()
-                        else:
-                            self._down_button.turn_off()
-
-    # Refresh button and its listener OK
-    def set_down_button(self, button):
-        assert (isinstance(button, (ButtonElement, type(None))))
-        if (self._down_button != button):
-            if (self._down_button != None):
-                self._down_button.remove_value_listener(self._down_button_value)
-            self._down_button = button
-            if (self._down_button != None):
-                assert isinstance(button, ButtonElement)
-                self._down_button.add_value_listener(self._down_button_value, identify_sender=True)
-            self._update_down_button()
-
-    # Handle page and scroll displacement in down direction
-    def _down_button_value(self, value, sender):
-        assert (self._down_button != None)
-        assert (value in range(128))
-        if self.is_enabled() and self._clip != None:
-            if value is not 0 or not sender.is_momentary():
-                if self.should_scroll():
-                    self.scroll_down()
-                else:
-                    self.page_down()
-                self._step_sequencer.update()
-
-    # Updates the UP button light OK 
-    def _update_up_button(self):
-        if self.is_enabled():
-            if self._up_button != None:
-                if self._clip == None:
-                    self._up_button.set_on_off_values("DefaultButton.Disabled", "DefaultButton.Disabled")
-                    self._up_button.turn_off()
-                else:
-                    self._up_button.set_on_off_values("StepSequencer.Octave.On", "StepSequencer.Octave.Off")
-                    if self.should_scroll():
-                        if self.can_scroll_up():
-                            self._up_button.turn_on()
-                        else:
-                            self._up_button.turn_off()
-                    else:
-                        if self.can_page_up():
-                            self._up_button.turn_on()
-                        else:
-                            self._up_button.turn_off()
-
-    # Refresh button and its listener OK
-    def set_up_button(self, button):
-        assert (isinstance(button, (ButtonElement, type(None))))
-        if self._up_button != button:
-            if self._up_button != None:
-                self._up_button.remove_value_listener(self._up_button_value)
-            self._up_button = button
-            if self._up_button != None:
-                assert isinstance(button, ButtonElement)
-                self._up_button.add_value_listener(self._up_button_value, identify_sender=True)
-            self._update_up_button()
-
-    # Handle page and scroll displacement in up direction
-    def _up_button_value(self, value, sender):
-        assert (self._up_button != None)
-        assert (value in range(128))
-        if self.is_enabled() and self._clip != None:
-            if value is not 0 or not sender.is_momentary(): #If NOTEON or is Toggle
-                #IF [mute_shift and NormalMode-> move(1)] [not mute_shift and MultinoteMode-> move(1)] ELSE move(12|16)_dependingIfdrumRack
-                if self.should_scroll():
-                    self.scroll_up()
-                else:
-                    self.page_up()
-                self._step_sequencer.update()
-
-    # Note Selector grid buttons listener (Normal Mode) and set view to the drum pad
-    # Mute all the entries for a note (a lane)
-    def note_offset_button_value(self, value, sender):
-        if self.is_enabled() and value > 0 and self._enable_offset_button: # Is NoteON and Normal Mode
-            if self._is_mute_shifted:
-                try:
-                    index = self._offset_buttons.index(sender)
-                except ValueError:
-                    index = -1
-                self._step_sequencer._note_editor.mute_lane(self._root_note + index)
-            else:
-                try:
-                    self._offset = self._offset_buttons.index(sender)
-                except ValueError:
-                    self._offset = -1
-                if self.is_drumrack and self._drum_group_device!=None:
-                    self._drum_group_device.view.selected_drum_pad = self._drum_group_device.drum_pads[self.selected_note]
-
-                self._step_sequencer._scale_updated()
-
-    def update(self):
-        if self.is_enabled():
-            self._step_sequencer._track_controller._do_implicit_arm(self._is_velocity_shifted and not self._step_sequencer._is_locked)
-            if self._is_velocity_shifted and not self._step_sequencer._is_locked:
-                self._control_surface.set_feedback_channels([11]) # WHY USE Channel 12 to play the notes???
-                self._control_surface._c_instance.set_feedback_velocity(int(self._control_surface._skin['Note.Feedback'])) # What is this???
-                self._was_velocity_shifted = True
-            elif self.is_drumrack and self._was_velocity_shifted:
-                self._was_velocity_shifted = False
-                self._offset = self._drum_group_device.view.selected_drum_pad.note - self._root_note
-                self._step_sequencer._scale_updated()
-            self._update_up_button()
-            self._update_down_button()
-            self._update_matrix()
-
-    def _update_matrix(self):
-        if self._enable_offset_button and self.is_enabled():
-            for i in range(len(self._offset_buttons)):
-                if self._clip == None:
-                    self._offset_buttons[i].set_light("DefaultButton.Disabled")
-                    self._offset_buttons[i].set_enabled(True)
-                else:
-                    note = self._root_note + i
-                    if self.is_drumrack:
-                        if self._drum_group_device.drum_pads[note].chains:
-                            self._offset_buttons[i].set_on_off_values("DrumGroup.PadSelected","DrumGroup.PadFilled")
-                        else:
-                            self._offset_buttons[i].set_on_off_values("DrumGroup.PadSelected", "DrumGroup.PadEmpty")
-                    else:
-                        if self._scale != None:
-                            if i % 12 == self._scale[0]:
-                                self._offset_buttons[i].set_on_off_values("StepSequencer.NoteSelector.Selected", "Note.Pads.Root")
-                            elif i % 12 == self._scale[2] or i % 12 == self._scale[4]:
-                                self._offset_buttons[i].set_on_off_values("StepSequencer.NoteSelector.Selected", "Note.Pads.Highlight")
-                            elif self._scale != None and i % 12 in self._scale:
-                                self._offset_buttons[i].set_on_off_values("StepSequencer.NoteSelector.Selected", "Note.Pads.InScale")
-                            else:
-                                self._offset_buttons[i].set_on_off_values("StepSequencer.NoteSelector.Selected", "Note.Pads.OutOfScale")
-                        else:
-                            self._offset_buttons[i].set_on_off_values("StepSequencer.NoteSelector.Selected", "Note.Pads.OutOfScale")
-
-                    if self._is_velocity_shifted and not self._step_sequencer._is_locked:
-                        #self._offset_buttons[i].force_next_send()
-                        #self._offset_buttons[i].turn_off()
-                        self._offset_buttons[i].set_enabled(False)
-                        self._offset_buttons[i].set_channel(11)
-                        self._offset_buttons[i].set_identifier(note)
-                    else:
-                        #self._offset_buttons[i].force_next_send()
-                        self._offset_buttons[i].set_enabled(True)
-                        self._offset_buttons[i].use_default_message()
-
-                        if self._playhead != None and self.note_is_playing(self._clip, self._note_cache, note, self._playhead):
-                            self._offset_buttons[i].set_on_off_values("StepSequencer.NoteSelector.Playing","StepSequencer.NoteSelector.Playing")
-
-                    if self.selected_note == note:
-                        if self._cache[i] != self._offset_buttons[i]._on_value or self._force:
-                            self._offset_buttons[i].turn_on()
-                            self._cache[i] = self._offset_buttons[i]._on_value
-                    else:
-                        if self._cache[i] != self._offset_buttons[i]._off_value or self._force:
-                            self._offset_buttons[i].turn_off()
-                            self._cache[i] = self._offset_buttons[i]._off_value
-            self._force = False
-
-    def set_enabled(self, enabled):
-        if enabled:
-            self._cache = [-1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1]
-        ControlSurfaceComponent.set_enabled(self, enabled)
-
-    def set_scale(self, scale, key=-1):
-        if key != -1:
-            self._key = key
-        self._scale = scale
-        # relativise scale
-        for i in range(len(self._scale)):
-            self._scale[i] = self._scale[i] - self._key
-
-    #Is the cursor in the current button range and contain a note
-    def note_is_playing(self, clip, note_cache, midi_note, playhead):
-        if clip != None and clip.is_playing and note_cache != None:
-            for note in note_cache:
-                note_key = note[0]
-                note_position = note[1]
-                note_length = note[2]
-                # note_velocity = note[3]
-                note_muted = note[4]
-                if midi_note == note_key and note_position <= playhead and note_position + note_length >= playhead and not note_muted:
-                    return True
-        return False
-
-    def scroll_down(self):
-        self.move(-1)
-
-    def scroll_up(self):
-        self.move(1)
-
-    def page_down(self):
-        if self.is_drumrack:
-            self.move(-16)
-        else:
-            self.move(-12)
-
-    def page_up(self):
-        if self.is_drumrack:
-            self.move(16)
-        else:
-            self.move(12)
-
-    def can_scroll_down(self):
-        return self.can_move(-1)
-
-    def can_scroll_up(self):
-        return self.can_move(1)
-
-    def can_page_down(self):
-        if self.is_drumrack:
-            return self.can_move(-16)
-        else:
-            return self.can_move(-12)
-
-    def can_page_up(self):
-        if self.is_drumrack:
-            return self.can_move(16)
-        else:
-            return self.can_move(12)
-        
-    #is in MIDI range
-    def can_move(self, steps):
-        return self.selected_note + steps >= 0 and self.selected_note + steps < 128
-
-    # Shifts the note selector up and down (12 if melodic instrument/16 if drumrack)
-    def move(self, steps):
-        if self.can_move(steps):
-            if self.is_diatonic:
-                # find the next note in scale in that direction
-                octave = 0
-                if steps > 0:
-                    while self._offset + steps >= 12:
-                        steps = steps - 12
-                        octave = octave + 1
-                    while steps != 0 and self._offset + steps not in self._scale:
-                        steps = steps + 1
-                        while self._offset + steps > 12:
-                            steps = steps - 12
-                            octave = octave + 1
-                elif steps < 0:
-                    while self._offset + steps < 0:
-                        steps = steps + 12
-                        octave = octave - 1
-                    while steps != 0 and self._offset + steps not in self._scale:
-                        steps = steps - 1
-                        if self._offset + steps < 0:
-                            steps = steps + 12
-                            octave = octave - 1
-                try:
-                    idx = self._scale.index(self._offset + steps)
-                except ValueError:
-                    idx = -1
-                self.set_selected_note(self._root_note + octave * 12 + self._scale[idx])
-            else:
-                self.set_selected_note(self._root_note + self._offset + steps)
-
-    def set_selected_note(self, selected_note):
-        if self.is_drumrack:
-            self._root_note = int((selected_note + 12) / 16 - 1) * 16 + 4
-            self._offset = (selected_note - self._root_note + 16) % 16
-        else:
-            self._root_note = int((selected_note - self._key) / 12) * 12 + self._key
-            self._offset = (selected_note + 12 - self._root_note) % 12
-        
-        self._step_sequencer._scale_updated()
-
-    def set_key(self, key):
-        self._key = key
-
-    def note_is_available(self, key): # deprecated???
-        if self.is_drumrack:
-            if self._drum_group_device.drum_pads[key].chains:
-                return True
-            else:
-                return False
-        else:
-            for note in self._scale:
-                if key % 12 == (self.root_note + note) % 12:
-                    return True
-                else:
-                    return False
-
-    def note_is_used(self, key):# deprecated???
-        if self._clip != None:
-            for note in self._clip_notes:
-                if note[0] == key:  # key: 0-127 MIDI note #
-                    return True
-        return False
-
-    @property
-    def selected_note(self):
-        return self._root_note + self._offset
-
-    def should_scroll(self):
-        return not self._is_mute_shifted and not self._enable_offset_button or self._is_mute_shifted and self._enable_offset_button    
-
-#Used in Normal mode (Not Multinote) to delete/copy/mute/change loops regions
-#Scrolls the regions in Multinote Mode 
-class LoopSelectorComponent(ControlSurfaceComponent):
-
-    def __init__(self, step_sequencer, buttons, control_surface):
-        ControlSurfaceComponent.__init__(self)
-        self._control_surface = control_surface
-        self.set_enabled(False)
-        self._step_sequencer = step_sequencer
-
-        self._clip = None # clip being played
-        self._notes = None # notes of the clip
-        self._playhead = None # contains the clip playing position
-
-        self._loop_end = 0
-        self._loop_start = 0
-
-        self._blocksize = 8  # number of notes per block -> how many steps are in a button (depending on quantization for note length variable)
-        self._block = 0  # currently selected block (button)
-        self._force = True # used to force a state change / message send
-
-        # used for loop selection
-        self._last_button_idx = -1
-        self._last_button_time = time.time()
-        self._loop_point1 = -1
-        self._loop_point2 = -1
-
-        self._cache = [-1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1] # Length=16
-
-        self._buttons = buttons
-        for button in self._buttons: # iterate 16 buttons of 4x4 lower right matrix section
-            assert isinstance(button, ButtonElement)
-            button.remove_value_listener(self._loop_button_value)
-            button.add_value_listener(self._loop_button_value, identify_sender=True)
-            
-    def disconnect(self):
-        self._top_buttons = None
-
-    @property
-    def _number_of_lines_per_note(self):
-        if self._mode == STEPSEQ_MODE_MULTINOTE:
-            return self._step_sequencer._number_of_lines_per_note
-        else:
-            return 1
-
-    def set_clip(self, clip):
-        self._clip = clip
-
-    @property
-    def _mode(self):
-        return self._step_sequencer._mode
-
-    def set_note_cache(self, note_cache):
-        self._note_cache = note_cache
-
-    def set_playhead(self, playhead):
-        self._playhead = playhead
-        self.update()
-
-    @property
-    def _is_mute_shifted(self):
-        return self._step_sequencer._is_mute_shifted
-
-    @property
-    def _is_velocity_shifted(self):
-        return self._step_sequencer._note_editor._is_velocity_shifted
-
-    @property
-    def _quantization(self):
-        return self._step_sequencer._quantization
-
-    @property
-    def block(self):
-        return self._block
-
-    def set_blocksize(self, blocksize):
-        self._blocksize = blocksize
-
-    def set_enabled(self, enabled):
-        self._force = True
-        ControlSurfaceComponent.set_enabled(self, enabled)
-
-    # Read Live's Clip loop values to LoopSelector Values OK
-    def _get_clip_loop(self): 
-        if self._clip != None:
-            self._loop_start = self._clip.loop_start
-            self._loop_end = self._clip.loop_end
-        else:
-            self._loop_start = 0
-            self._loop_end = 0
-
-    # Write LoopSelector Values to Live's Clip loop values (loop and marker) OK
-    def set_clip_loop(self, start, end): 
-        if self._clip != None:
-            self._loop_end = end
-            self._loop_start = start
-            if self._loop_start >= self._clip.loop_end:
-                self._clip.loop_end = self._loop_end
-                self._clip.loop_start = self._loop_start
-                self._clip.end_marker = self._loop_end
-                self._clip.start_marker = self._loop_start
-            else:
-                self._clip.loop_start = self._loop_start
-                self._clip.loop_end = self._loop_end
-                self._clip.start_marker = self._loop_start
-                self._clip.end_marker = self._loop_end
-            self.update()
-
-    #LoopSelector listener OK
-    def _loop_button_value(self, value, sender): #Allows to make selection by hold and pressing marker buttons
-        # Selects simple page by double click on region button
-        # Allows to mute and delete notes in a range
-        # Allows to duplicate a range forwards (to empty regions)
-        if self.is_enabled():
-            idx = self._buttons.index(sender)
-            if value > 0:#This allow to setup loop range by pressing two buttons at a time [Start,End]
-                if self._loop_point1 == -1:
-                    self._loop_point1 = idx
-                elif self._loop_point2 == -1:
-                    self._loop_point2 = idx
-        
-            #Button released
-            elif self._loop_point1 != -1: 
-                setloop = self._loop_point2 != -1 # two buttons pressed
-                if self._loop_point2 == -1:
-                    self._loop_point2 = idx # _loop_point1 = _loop_point2 
-                    if self._last_button_idx == idx and (time.time() - self._last_button_time) < 0.25: #Double clic set loop subsection (depending on quantization)
-                        setloop = True
-                        self._last_button_time = time.time()
-                        self._last_button_idx = -1
-
-                if self._loop_point1 != -1 and self._loop_point2 != -1:
-                    
-                    start= min(self._loop_point1, self._loop_point2)
-                    end= max(self._loop_point1, self._loop_point2) + 1
-                    
-                    self._block = start
-                    if setloop:
-                        if self._is_mute_shifted:
-                            if self._is_velocity_shifted:
-                                self._mute_notes_in_range(start * self._blocksize * self._quantization, end * self._blocksize * self._quantization)
-                            else:
-                                self._delete_notes_in_range(start * self._blocksize * self._quantization, end * self._blocksize * self._quantization)
-                        else:
-                            if self._is_velocity_shifted: # FIX, see if can copy backwards
-                                self._extend_clip_content(start * self._blocksize * self._quantization, self._loop_end, end * self._blocksize * self._quantization)
-                            self.set_clip_loop(start * self._blocksize * self._quantization, end * self._blocksize * self._quantization)
-
-                    self._step_sequencer.set_page(self._block) # set sequencer focus
-                    self._loop_point1 = -1
-                    self._loop_point2 = -1
-                    self.update()
-                self._last_button_time = time.time()
-                self._last_button_idx = idx
-
-    #Index check for page boundaries scroll OK
-    def can_scroll(self, blocks):
-        if self._clip != None:
-            if (blocks + self._block) < 0:
-                return False
-            if (blocks + self._block) * 8 * self._quantization * self._number_of_lines_per_note < self._clip.loop_start:
-                return False
-            if (blocks + self._block + 1) * 8 * self._quantization * self._number_of_lines_per_note > self._clip.loop_end:
-                return False
-            return True
-        return False
-
-    # Does the actual scroll OK
-    def scroll(self, blocks):
-        if self._clip != None and self.can_scroll(blocks):
-            self._block = blocks + self._block
-            self._step_sequencer.set_page(self._block)
-
-    # Iterates refreshing all loop selector buttons (called from playing position listener) OK
-    def update(self):
-        if self.is_enabled():
-            self._get_clip_loop() # gets the loop start/end values from the clip -> self._loop_start & self._loop_end
-            i = 0
-            for button in self._buttons: # iterate 16 buttons of 4x4 lower right matrix section
-                if self._clip == None: #Disable/turn off all buttons
-                    button.set_on_off_values("DefaultButton.Disabled","DefaultButton.Disabled")
-                    if self._cache[i] != button._off_value:
-                        button.turn_off()
-                        self._cache[i] = button._off_value
-                else:
-                    #is the button in loop range
-                    in_loop = (i * self._blocksize * self._quantization < self._loop_end) and (i * self._blocksize * self._quantization >= self._loop_start)
-                    #is the playing position is inside the block represented by the button
-                    playing = self._playhead != None and self._playhead >= i * self._blocksize * self._quantization and self._playhead < (i + 1) * self._blocksize * self._quantization
-                    #is this block selected (green)
-                    selected = i == self.block
-                    if in_loop:
-                        if playing:
-                            if selected:
-                                self._cache[i] = "StepSequencer.LoopSelector.SelectedPlaying"
-                            else:
-                                self._cache[i] = "StepSequencer.LoopSelector.Playing"
-                        else:
-                            if selected:
-                                self._cache[i] = "StepSequencer.LoopSelector.Selected"
-                            else:
-                                self._cache[i] = "StepSequencer.LoopSelector.InLoop"    
-                    else:
-                        if playing:
-                            if selected:
-                                self._cache[i] = "StepSequencer.LoopSelector.SelectedPlaying"
-                            else:
-                                self._cache[i] = "StepSequencer.LoopSelector.Playing"
-                        else:
-                            if selected:
-                                self._cache[i] = "StepSequencer.LoopSelector.Selected"
-                            else:
-                                self._cache[i] = "DefaultButton.Disabled"
-
-                    if self._cache[i] != button._on_value or self._force:#Enable/turn on all buttons
-                        button.set_on_off_values(self._cache[i], self._cache[i])
-                        button.turn_on()
-                i = i + 1
-            self._force = False
-
-    #Make a copy of the current loop to the next N empty blocks OK
-    def _extend_clip_content(self, loop_start, old_loop_end, new_loop_end):
-        if(self._no_notes_in_range(old_loop_end, new_loop_end, True)):
-            clip_looping_length = 0
-            if(old_loop_end > 1):
-                power = 1
-                while(power * 2 < old_loop_end):
-                    power *= 2
-                clip_looping_length = (power)
-            clone_length = new_loop_end - old_loop_end
-            if(clip_looping_length > 0):
-                clone_start_point = (old_loop_end % clip_looping_length)
-            else:
-                clone_start_point = 0
-            self._copy_notes_in_range(clone_start_point, clone_start_point + clone_length, old_loop_end)
-
-    #Does the note by note copy OK
-    def _copy_notes_in_range(self, start, end, new_start):
-        new_notes = list(self._note_cache)
-        # for i in range()
-        for note in new_notes:
-            if note[1] >= start and note[1] < end:
-                new_notes.append([note[0], note[1] + new_start - start, note[2], note[3], note[4]])
-        self._clip.select_all_notes()
-        self._clip.replace_selected_notes(tuple(new_notes))
-    
-    #Checks if a range is empty OK
-    def _no_notes_in_range(self, start, end, or_after):
-        for note in list(self._note_cache):
-            if note[1] >= start and (note[1] < end or or_after):
-                return(False)
-        return(True)
-
-    #Deletes a block of notes OK
-    def _delete_notes_in_range(self, start, end):
-        new_notes = list()
-        for note in list(self._note_cache):
-            if note[1] < start or note[1] >= end:
-                new_notes.append(note)
-        self._clip.select_all_notes()
-        self._clip.replace_selected_notes(tuple(new_notes))
-
-    #Mutes a block of notes OK
-    def _mute_notes_in_range(self, start, end):
-        new_notes = list()
-        for note in list(self._note_cache): #Note -> tuple containing pitch, time, duration, velocity, and mute
-            if note[1] < start or note[1] >= end: #Note time
-                new_notes.append(note)
-            else:
-                new_notes.append([note[0], note[1], note[2], note[3], not note[4]]) # Negate mute state
-        self._clip.select_all_notes()
-        self._clip.replace_selected_notes(tuple(new_notes))
+#Scrolls the regions in Multinote Mode
 
 
 class StepSequencerComponent(CompoundComponent):
@@ -780,12 +125,13 @@ class StepSequencerComponent(CompoundComponent):
 
     # Set 4x4 lower right matrix section that manages the loop range OK
     def _set_loop_selector(self):
-        self._loop_selector = self.register_component(LoopSelectorComponent(self, [
+        self._loop_selector = self.register_component(
+            LoopSelectorComponent(self, [
             self._matrix.get_button(4, 4), self._matrix.get_button(5, 4), self._matrix.get_button(6, 4), self._matrix.get_button(7, 4),
             self._matrix.get_button(4, 5), self._matrix.get_button(5, 5), self._matrix.get_button(6, 5), self._matrix.get_button(7, 5),
             self._matrix.get_button(4, 6), self._matrix.get_button(5, 6), self._matrix.get_button(6, 6), self._matrix.get_button(7, 6),
             self._matrix.get_button(4, 7), self._matrix.get_button(5, 7), self._matrix.get_button(6, 7), self._matrix.get_button(7, 7)],
-            self._control_surface)
+                                  self._control_surface)
             )
             
     #Allow to manipulate the LP grid and Live's Clip notes (add/del, velocity, mute, etc)
@@ -797,12 +143,13 @@ class StepSequencerComponent(CompoundComponent):
 
     #Set 4x4 lower left matrix section that allows note selection in Normal Mode
     def _set_note_selector(self):
-        self._note_selector = self.register_component(NoteSelectorComponent(self, [
+        self._note_selector = self.register_component(
+            NoteSelectorComponent(self, [
             self._matrix.get_button(0, 7), self._matrix.get_button(1, 7), self._matrix.get_button(2, 7), self._matrix.get_button(3, 7),
             self._matrix.get_button(0, 6), self._matrix.get_button(1, 6), self._matrix.get_button(2, 6), self._matrix.get_button(3, 6),
             self._matrix.get_button(0, 5), self._matrix.get_button(1, 5), self._matrix.get_button(2, 5), self._matrix.get_button(3, 5),
             self._matrix.get_button(0, 4), self._matrix.get_button(1, 4), self._matrix.get_button(2, 4), self._matrix.get_button(3, 4)],
-            self._control_surface)
+                                  self._control_surface)
             )
         self._note_selector.set_up_button(self._side_buttons[4])#Stop
         self._note_selector.set_down_button(self._side_buttons[5])#Trk On
